requirements.txt
fastapi==0.112.0
uvicorn==0.30.5
SQLAlchemy==2.0.32
pydantic==2.8.2
pydantic_core==2.20.1
python-dotenv==1.0.1
psycopg2-binary==2.9.9
psycopg2-binary==2.9.9
python-jose==3.3.0